Declarations containning all the control signals and input and ouputs of basic blocks
All the control signals are of type integer. Instruction is a string. The memory is  a dictionary initialised by a file 1.mc. All registers are initialised to zero in decimal format except for sp.
All other memory can be assummed to be zero. All the signals can be identified with the diagram sent in the group